fix: return active style to categories button
