--- src/constants/productCategories/productCategories.ts
@@ -1,32 +1,0 @@
-export const CATEGORIES_OF_PRODUCTS = [
-  {
-    className: 'products-nav-item products-nav-item_active',
-    data: 'All Categories',
-    id: Math.random(),
-  },
-
-  {
-    data: 'Free',
-    id: Math.random(),
-  },
-
-  {
-    data: 'Programming',
-    id: Math.random(),
-  },
-
-  {
-    data: 'Data analysis',
-    id: Math.random(),
-  },
-
-  {
-    data: 'Design',
-    id: Math.random(),
-  },
-
-  {
-    data: 'Management',
-    id: Math.random(),
-  },
-];
