+++ src/constants/productsConstant/productsConstants.ts
@@ -10,3 +10,36 @@ export const PRODUCTS_SORT_DATA = [
     id: Math.random(),
   },
 ];
+
+export const CATEGORIES_OF_PRODUCTS = [
+  {
+    className: 'products-nav-item_active',
+    data: 'All Categories',
+    id: Math.random(),
+  },
+
+  {
+    data: 'Free',
+    id: Math.random(),
+  },
+
+  {
+    data: 'Programming',
+    id: Math.random(),
+  },
+
+  {
+    data: 'Data analysis',
+    id: Math.random(),
+  },
+
+  {
+    data: 'Design',
+    id: Math.random(),
+  },
+
+  {
+    data: 'Management',
+    id: Math.random(),
+  },
+];
