+++ src/constants/scss/extentds.scss
@@ -7,7 +7,7 @@
     text-align: center;
     border: none;
     background-color: inherit;
-    padding: 5px 10px;
+    padding: 10px;
 }
 
 %mainButtonGradient {
